Fix calling torch.compile inside of a `__torch_dispatch__`
We should only bail out in the following situation:
```py
with torch_dispatch_mode()
   torch.compile(f)(x)
```
However, before this PR, we are also bailing out in the following
situation:
```py
with torch_dispatch_mode()
   no_compile_f(x)
```
where the torch_dispatch_mode's __torch_dispatch__ calls torch.compile.

In order to only error out in the first situation, the check we should
be doing is "are there any non-infra modes on the current torch_dispatch
stack". That is what this PR adds.

Test Plan:
- new test

ghstack-source-id: 3c6abf5be4aeabedf0af6fffc57b5d71596ed1af
Pull Request resolved: https://github.com/pytorch/pytorch/pull/148712